fix: circular import (move helper functions)
